more cleanup in cvs/git
